+++ src/cam_touptek.cpp
@@ -870,13 +870,13 @@ bool CameraToupTek::GetCoolerStatus(bool *on, double *setpoint, double *power, d
         err = true;
 
     if (m_cam.GetOption(TOUPCAM_OPTION_TECTARGET, &targ))
-        *setpoint = static_cast<double>(targ / 10.0);
+        *setpoint = targ / 10.0;
     else
         err = true;
 
     if (m_cam.GetOption(TOUPCAM_OPTION_TEC_VOLTAGE, &vcur) && m_cam.GetOption(TOUPCAM_OPTION_TEC_VOLTAGE_MAX, &vmax) &&
         vmax > 0)
-        *power = static_cast<double>(vcur * 100.0 / vmax);
+        *power = vcur * 100.0 / vmax;
     else
         err = true;
 
@@ -897,7 +897,7 @@ bool CameraToupTek::GetSensorTemperature(double *temperature)
         return true;
     }
 
-    *temperature = static_cast<double>(val / 10.0);
+    *temperature = val / 10.0;
     return false;
 }
